move PinPolicy to common
