+++ src/RutokenPkcs11Interop/RutokenPkcs11Interop/Common/PinPolicy.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RutokenPkcs11Interop.HighLevelAPI
+namespace RutokenPkcs11Interop.Common
 {
     public class PinPolicy
     {
