+++ src/RutokenPkcs11Interop/RutokenPkcs11Interop/RutokenPkcs11Interop.csproj
@@ -89,7 +89,7 @@
     <Compile Include="HighLevelAPI41\MechanismParams\CkGostR3410_12_256_DeriveParams.cs" />
     <Compile Include="HighLevelAPI41\MechanismParams\CkGostR3410_12_DeriveParams.cs" />
     <Compile Include="HighLevelAPI41\Pkcs11Extensions.cs" />
-    <Compile Include="HighLevelAPI\PinPolicy.cs" />
+    <Compile Include="Common\PinPolicy.cs" />
     <Compile Include="HighLevelAPI\Pkcs7VerificationResult.cs" />
     <Compile Include="HighLevelAPI41\RutokenInitParam.cs" />
     <Compile Include="HighLevelAPI41\SessionExtensions.cs" />
