+++ src/RutokenPkcs11Interop/RutokenPkcs11InteropTests/HighLevelAPI/_HL_35_PinPolicyTest.cs
@@ -2,6 +2,7 @@
 using Net.Pkcs11Interop.Common;
 using NUnit.Framework;
 using Net.Pkcs11Interop.HighLevelAPI;
+using RutokenPkcs11Interop.Common;
 using RutokenPkcs11Interop.HighLevelAPI;
 
 namespace RutokenPkcs11InteropTests.HighLevelAPI
